LPD-55581 - Move ReactFieldBase under /api and update references
