+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/Checkbox/Checkbox.tsx
@@ -6,7 +6,7 @@
 import {ClayInput} from '@clayui/form';
 import React from 'react';
 
-import FieldBase from '../FieldBase/ReactFieldBase.es';
+import FieldBase from '../js/api/FieldBase/ReactFieldBase';
 import CheckboxLocalizedObjectField, {
 	IProps as ICheckboxLocalizedObjectFieldProps,
 } from '../localizedObjectFields/CheckboxLocalizedObjectField';
