+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/Options/Options.es.js
@@ -13,7 +13,7 @@ import React, {useEffect, useRef, useState} from 'react';
 import {DndProvider} from 'react-dnd';
 import {HTML5Backend} from 'react-dnd-html5-backend';
 
-import FieldBase from '../FieldBase/ReactFieldBase.es';
+import FieldBase from '../js/api/FieldBase/ReactFieldBase';
 import OptionFieldKeyValue from '../OptionFieldKeyValue/OptionFieldKeyValue';
 import DnD from './DnD.es';
 import DragPreview from './DragPreview.es';
