+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/Paragraph/Paragraph.es.js
@@ -5,7 +5,7 @@
 
 import React from 'react';
 
-import FieldBase from '../FieldBase/ReactFieldBase.es';
+import FieldBase from '../js/api/FieldBase/ReactFieldBase';
 
 const Paragraph = ({name, text, ...otherProps}) => (
 	<FieldBase {...otherProps} name={name} text={text}>
