+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/SearchLocation/SearchLocation.es.js
@@ -8,7 +8,7 @@ import {SettingsContext, useFormState} from 'data-engine-js-components-web';
 import {openToast} from 'frontend-js-components-web';
 import React, {useEffect, useState} from 'react';
 
-import FieldBase from '../FieldBase/ReactFieldBase.es';
+import FieldBase from '../js/api/FieldBase/ReactFieldBase';
 import usePlaces from './usePlaces.es';
 
 const parse = (value, defaultValue) => {
