+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/Select/Select.tsx
@@ -6,7 +6,7 @@
 import {useFormState} from 'data-engine-js-components-web';
 import React, {useMemo} from 'react';
 
-import FieldBase from '../FieldBase/ReactFieldBase.es';
+import FieldBase from '../js/api/FieldBase/ReactFieldBase';
 import SelectLocalizedObjectField, {
 	SelectLocalizedObjectFieldProps,
 } from '../localizedObjectFields/SelectLocalizedObjectField';
