+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/Separator/Separator.es.js
@@ -5,7 +5,7 @@
 
 import React, {useEffect, useRef} from 'react';
 
-import FieldBase from '../FieldBase/ReactFieldBase.es';
+import FieldBase from '../js/api/FieldBase/ReactFieldBase';
 
 const Separator = ({name, style, ...otherProps}) => {
 	const elRef = useRef(null);
