+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/index.ts
@@ -22,7 +22,6 @@ import ColorPicker from './ColorPicker/ColorPicker.es';
 // @ts-ignore
 
 import DocumentLibrary from './DocumentLibrary/DocumentLibrary.es';
-import ReactFieldBase from './FieldBase/ReactFieldBase.es';
 
 // @ts-ignore
