+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/js/api/FieldBase/ReactFieldBase.d.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {LocalizedValue} from '../types';
+type LocalizedValue<T> = Liferay.Language.LocalizedValue<T>;
 
 export default function FieldBase(props: IFieldBase): JSX.Element;
