+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/js/api/api.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default as ReactFieldBase} from '../../FieldBase/ReactFieldBase.es';
+export {default as ReactFieldBase} from './FieldBase/ReactFieldBase';
