+++ modules/apps/fragment/fragment-impl/src/main/resources/META-INF/resources/js/api/api.ts
@@ -3,6 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+// @ts-ignore
+
 export {LocalizationSelect} from './LocalizationSelect';
 export {getOrCreateTranslationInput} from './getOrCreateTranslationInput';
 export {
