+++ modules/apps/fragment/fragment-impl/src/main/resources/META-INF/resources/js/api/registerUnlocalizedInput.ts
